+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Version 0.0.5
+
+- Sort order for sensors
 - Support for BH1750FVI Ambient Light Sensor
 - Removed persistant WiFi settings and improved stability
 - WebServer logging to Serial